requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==0.2.28
langchain==0.3.3
langchain-google-genai
langsmith==0.1.131
google-api-python-client==2.149.0
google-auth==2.35.0
notion-client==2.2.1
pyairtable==2.3.3
sendgrid==6.11.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
